% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_df_hg19}
\alias{genes_df_hg19}
\title{an hg19-based data.frame with gene addresses}
\format{
data.frame
}
\usage{
genes_df_hg19
}
\description{
an hg19-based data.frame with gene addresses
}
\note{
Faster response than GRanges/genes on TxDb.  \code{genes_df_hg19[sym,]} gives one row with seqnames, start, end, etc.
}
\keyword{datasets}
