% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hits_near_ormdl3_trunc10}
\alias{hits_near_ormdl3_trunc10}
\title{a GRanges derived from the EBI/EMBL GWAS catalog of 1 Nov 2020}
\format{
GenomicRanges GRanges instance
}
\usage{
hits_near_ormdl3_trunc10
}
\description{
a GRanges derived from the EBI/EMBL GWAS catalog of 1 Nov 2020
}
\note{
The 'value' field is the result of \code{pmin(hits_near_ormdl3_trunc10$PVALUE_MLOG,10)},
which is convenient to see variation in significance, which would otherwise be crushed
down owing to mlog values in the 100s.
}
\keyword{datasets}
