% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lamb_main_20201101}
\alias{lamb_main_20201101}
\title{the main table of Lambert et al., Cell 2018: "The human transcription factors"}
\format{
tibble
}
\usage{
lamb_main_20201101
}
\description{
the main table of Lambert et al., Cell 2018: "The human transcription factors"
}
\keyword{datasets}
