% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcov.R
\name{linetrack_covg_by_symbol}
\alias{linetrack_covg_by_symbol}
\title{produce a TnT::LineTrack with coverage information for a selected gene+radius}
\usage{
linetrack_covg_by_symbol(sym, bamf, radius = 0, ...)
}
\arguments{
\item{sym}{a gene symbol}

\item{bamf}{a path to indexed BAM file}

\item{radius}{numeric(1) defaulting to zero, upstream/downstream extra reach}

\item{\dots}{passed to LineTrack, so can include color, etc.}
}
\description{
produce a TnT::LineTrack with coverage information for a selected gene+radius
}
\examples{
if (requireNamespace("RNAseqData.HNRNPC.bam.chr14")) {
  f1 = RNAseqData.HNRNPC.bam.chr14::RNAseqData.HNRNPC.bam.chr14_BAMFILES[1]
  linetrack_covg_by_symbol("HNRNPC", f1, color="gray")
}
}
