% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfuns.R
\name{reset_color}
\alias{reset_color}
\title{set TnT glyph color data using EnsDb metadata}
\usage{
reset_color(view, ttcolumn = "TXBIOTYPE")
}
\arguments{
\item{view}{a TnT TxTrack instance}

\item{ttcolumn}{column name that is already present in tooltip metadata}
}
\value{
TxTrack
}
\description{
set TnT glyph color data using EnsDb metadata
}
