% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfuns.R
\name{reset_display_label}
\alias{reset_display_label}
\title{set TnT display labeling for a transcript set}
\usage{
reset_display_label(view, cols2use = c("SYMBOL", "TXBIOTYPE"))
}
\arguments{
\item{view}{a TnT TxTrack instance}

\item{cols2use}{column names already present in tooltip metadata}
}
\value{
TxTrack display label reset using TnT::strandlabel
}
\description{
set TnT display labeling for a transcript set
}
