% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfuns.R
\name{reset_tooltip}
\alias{reset_tooltip}
\title{set TnT tooltip data using EnsDb metadata}
\usage{
reset_tooltip(
  view,
  ensobj = EnsDb.Hsapiens.v79::EnsDb.Hsapiens.v79,
  cols2use = c("GENEID", "SYMBOL", "TXBIOTYPE")
)
}
\arguments{
\item{view}{a TnT TxTrack instance}

\item{ensobj}{an EnsDb instance}

\item{cols2use}{character() vector with elements drawn from result of \code{columns(ensobj)}}
}
\value{
TxTrack revised with specific tooltip data
}
\description{
set TnT tooltip data using EnsDb metadata
}
