% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tnt_gt_sym}
\alias{tnt_gt_sym}
\title{a TnT gene track with gene symbols as display_id, hg19}
\format{
TnT GeneTrack
}
\usage{
tnt_gt_sym
}
\description{
a TnT gene track with gene symbols as display_id, hg19
}
\keyword{datasets}
