% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tnt_txtr_14}
\alias{tnt_txtr_14}
\title{a TnT transcript track confined to chr14, hg19}
\format{
TnT GeneTrack
}
\usage{
tnt_txtr_14
}
\description{
a TnT transcript track confined to chr14, hg19
}
\keyword{datasets}
