% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txdata_near_ormdl3}
\alias{txdata_near_ormdl3}
\title{a transcript catalog on chr17}
\format{
GenomicRanges GRanges instance, generated using biovizBase::crunch on EnsDB.Hsapiens.v79.
The resulting GRanges was then limited to a 5Mb window starting at 38e6.  The mcols column
'type' is set to 'exon'.
}
\usage{
txdata_near_ormdl3
}
\description{
a transcript catalog on chr17
}
\keyword{datasets}
